Agregando prueba en cypress de concursos para equipos
